+++ packages/client/base/src/services/api/BaseCrossmintService.ts
@@ -1,33 +1,22 @@
 import { CrossmintServiceError } from "@/error";
 import { SDKLogger } from "@/utils/SDKLogger";
 
-import { validateAPIKey } from "@crossmint/common-sdk-base";
-
-import { CROSSMINT_DEV_URL, CROSSMINT_PROD_URL, CROSSMINT_STG_URL } from "../../consts";
 import { APIErrorService } from "./APIErrorService";
+import { getCrossmintBaseUrl } from "./helpers/getCrossmintBaseUrl";
 
 export abstract class BaseCrossmintService {
     public crossmintAPIHeaders: Record<string, string>;
     crossmintBaseUrl: string;
     protected abstract apiErrorService: APIErrorService<string>;
     protected abstract logger: SDKLogger;
-    private static urlMap: Record<string, string> = {
-        development: CROSSMINT_DEV_URL,
-        staging: CROSSMINT_STG_URL,
-        production: CROSSMINT_PROD_URL,
-    };
 
     constructor(apiKey: string) {
-        const result = validateAPIKey(apiKey);
-        if (!result.isValid) {
-            throw new Error("API key invalid");
-        }
+        this.crossmintBaseUrl = getCrossmintBaseUrl(apiKey);
         this.crossmintAPIHeaders = {
             accept: "application/json",
             "content-type": "application/json",
             "x-api-key": apiKey,
         };
-        this.crossmintBaseUrl = this.getUrlFromEnv(result.environment);
     }
 
     protected async fetchCrossmintAPI(
@@ -70,14 +59,5 @@ export abstract class BaseCrossmintService {
             },
             { endpoint: urlPath }
         );
-    }
-
-    protected getUrlFromEnv(environment: string) {
-        const url = BaseCrossmintService.urlMap[environment];
-        if (!url) {
-            console.log(" CrossmintService.urlMap: ", BaseCrossmintService.urlMap);
-            throw new Error(`URL not found for environment: ${environment}`);
-        }
-        return url;
     }
 }
