+++ packages/client/base/src/services/api/index.ts
@@ -1,2 +1,3 @@
 export * from "./APIErrorService";
 export * from "./BaseCrossmintService";
+export * from "./helpers";
