--- packages/client/ui/react-ui/src/components/auth/index.ts
@@ -1,1 +1,0 @@
-export * from "./AuthModal";
