+++ packages/client/ui/react-ui/src/components/index.ts
@@ -3,4 +3,3 @@ export * from "./CrossmintNFTDetail";
 
 export * from "./embed";
 export * from "./hosted";
-export * from "./auth";
