+++ packages/client/ui/react-ui/src/utils/constants.ts
@@ -1,2 +1,1 @@
-export const AUTH_SERVICE = "AUTH_SDK";
 export const SESSION_PREFIX = "crossmint-session";
